+++ src/App.tsx
@@ -1,5 +1,5 @@
 import React, { FC } from 'react'
-import PostComponent from './containers/Post'
+import TweetForm from './containers/TweetForm'
 import { Header, Icon } from 'semantic-ui-react'
 
 const App: FC = () => (
@@ -13,7 +13,7 @@ const App: FC = () => (
         Reactter
       </Header>
     </header>
-    <PostComponent />
+    <TweetForm />
   </>
 )
